pyproject.toml
[tool.poetry]
name = "crewai_gcp"
version = "0.1.0"
description = "crewai-gcp using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"  # Specify the Python version you are using
crewai = "0.67.1"
crewai-tools = "0.12.0"
pyyaml = "6.0.2"
python-dotenv = "1.0.1"
requests = "2.32.3"
google-auth-oauthlib = "1.2.1"
google-api-python-client = "2.160.0"


[tool.poetry.scripts]
crewai_gcp = "crewai_gcp.main:run"
train = "crewai_gcp.main:train"
replay = "crewai_gcp.main:replay"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
